requirements.txt
blinker==1.9.0
click==8.2.0
colorama==0.4.6
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyMySQL==1.1.0
python-dotenv==1.1.0
SQLAlchemy==2.0.23
typing_extensions==4.13.2
Werkzeug==3.1.3
